A revised, beta-targeted, approach to PR 96970.
This carries on with some of the ideas from PR 96970, but generalizes them
slightly to try to narrow the cases where the error fires. (It also sidesteps
the whole question of trying to deal with lifetime resolution during ast
lowering by just looking at names.)

More specifically: This tracks the depth of Return Position Opaque Ty (aka
Return Position Impl Trait aka RPIT). As it descends the AST, each time it does
lower_poly_trait_ref, it tracks what lifetimes are introduced there, and what
depth they were introduced at.

(As I write this, I realize that I probably didn't handle shadowed lifetimes
quite right; see below.)

Anyway, when it hits a use of a lifetime, it checks the depth. If the depths
don't match and we're in a nested RPIT, then it errors. (Compare this to PR
96970, which just errors unconditionally in the nested RPIT, without checking if
the depths match and thus it's a `for<'a>` that was introduced at the current
level as well, and thus should not be a problem.)

Regarding shadowing mistake: my code probably generates an error in a situation
like `impl for<'a> Trait1<Assoc=impl for<'a> Trait2<'a>>`, or something along
those lines. But that should be easy to fix.)

A follow-up commit will include the blessed test output, so you can see the
effect of this and compare it against PR 96970.

(I did not put in any new unit tests, but obviously we should make those too.)