Use `owned_slice` instead of `owning_ref` to store metadata.
`owning_ref` of a `Box<T>` has soundness problems because it stores
pointers to a box that is then invalidated. Instead, use the safer
`owned_slice`, which presents a more specialized abstraction.
This doesn't change the general layout or indirection of the structure,
but only the datastructure used as a wrapper.