Don't internalize symbols used by `global_asm`
Symbols that are only reachable from the current CGU have their
visibility changed to internal to allow LLVM to better optimize them.

However this doesn't work for symbols that are used in `global_asm!`:
the symbol name is inserted into the assembly code as text, which LLVM
is not aware of. In particular, LLVM may delete the symbol in the
MergeFunctions pass if it can't see any other uses of it.

The solution here is to ensure that symbols used by `global_asm!` are
always marked with external linkage.

Fixes #96797